refactor(decorators): remove deprecated decorators
BREAKING CHANGE
- use `@CollectionProperty(opts)`  instead of `@SortedSet`, `@TypedSet`, `@TypedArray`
- use `@Property({mapper: MapperForType})` instead of @CustomMapper(mapper)
