+++ src/decorator/impl/index.ts
@@ -1,9 +1,8 @@
 // collection
-export * from './collection/sorted-set.decorator'
-export * from './collection/typed-set.decorator'
-export * from './collection/typed-array.decorator'
+export * from './collection/collection-property.decorator'
+
 // date
-export * from './date/date.decorator'
+export * from './date/date-property.decorator'
 // index
 export * from './index/secondary-index'
 export * from './index/gsi-partition-key.decorator'
@@ -14,11 +13,13 @@ export * from './index/index-type.enum'
 export * from './key/partition-key.decorator'
 export * from './key/partition-key-uuid.decorator'
 export * from './key/sort-key.decorator'
-// mapper
-export * from './custom-mapper/custom-mapper.decorator'
+
 // model
 export * from './model/model.decorator'
 // property
 export * from './property/property.decorator'
 // transient
 export * from './transient/transient.decorator'
+
+
+
