+++ src/dynamo/expression/functions/attribute-names.const.ts
@@ -1,0 +1,2 @@
+export const NESTED_ATTR_PATH_CAPTURED_REGEX = /([a-z0-9A-Z_]+)(?:\[(\d+)])?\.?/g
+export const NESTED_ATTR_PATH_REGEX = /^.+((\[(\d+)])|(\.)).*$/
