+++ src/dynamo/expression/functions/attribute-names.function.ts
@@ -1,7 +1,7 @@
-import { Metadata } from '../../../decorator/metadata'
+import { Metadata } from '../../../decorator/metadata/metadata'
+import { NESTED_ATTR_PATH_CAPTURED_REGEX, NESTED_ATTR_PATH_REGEX } from './attribute-names.const'
 
-export const NESTED_ATTR_PATH_CAPTURED_REGEX = /([a-z0-9A-Z_]+)(?:\[(\d+)])?\.?/g
-export const NESTED_ATTR_PATH_REGEX = /^.+((\[(\d+)])|(\.)).*$/
+
 
 // problem: we only get the metadata from the last property -> but we need it for all properties in the chain (prop1.prop2.prop3)
 export function resolveAttributeNames(
