+++ test/models/nested-complex.model.ts
@@ -1,5 +1,4 @@
-import { CollectionProperty } from '../../src/decorator/impl/collection/collection-property.decorator'
-import { Model } from '../../src/dynamo-easy'
+import { CollectionProperty, Model } from '../../src/dynamo-easy'
 
 @Model()
 export class NestedComplexModel {
